fix: refactor a few more tests
